+++ crates/atuin/Cargo.toml
@@ -65,6 +65,12 @@ vendored-tls = ["atuin-client?/vendored-tls", "atuin-lab-share?/vendored-tls"]
 # axoasset's rustls/aws-lc-rs reqwest backend, which is only avoided by the
 # [patch.crates-io] axoasset fork when building from this repo.
 self-update = ["dep:axoupdater"]
+profiling-traced = [
+    "dep:tracing-opentelemetry",
+    "dep:opentelemetry",
+    "dep:opentelemetry_sdk",
+    "dep:opentelemetry-otlp",
+]
 
 [dependencies]
 atuin-ai = { path = "../atuin-ai", version = "18.20.0-beta.3", optional = true, default-features = false }
@@ -110,10 +116,10 @@ ratatui = { workspace = true }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 tracing-appender = "0.2"
-tracing-opentelemetry = "0.33"
-opentelemetry = "0.32"
-opentelemetry_sdk = "0.32"
-opentelemetry-otlp = "0.32"
+tracing-opentelemetry = { version = "0.33", optional = true }
+opentelemetry = { version = "0.32", optional = true }
+opentelemetry_sdk = { version = "0.32", optional = true }
+opentelemetry-otlp = { version = "0.32", optional = true }
 uuid = { workspace = true }
 sysinfo = "0.30.7"
 regex = "1.10.5"
